+++ Tests/Property.test.ps1
@@ -18,19 +18,19 @@ param
 
 . .\Shared.ps1
 
-assert ($Param1 -eq $BuildFile)
-assert ($Param2 -eq $env:USERNAME)
+equals $Param1 $BuildFile
+equals $Param2 $env:USERNAME
 
-assert ((property BuildFile) -eq $BuildFile)
-assert ((property UserName) -eq $env:USERNAME)
+equals (property BuildFile) $BuildFile
+equals (property UserName) $env:USERNAME
 
 $MissingProperty = property MissingProperty 42
-assert ($MissingProperty -eq 42)
+equals $MissingProperty 42
 
 # Null value is treated as missing, too
 $MissingNullProperty = $null
 $MissingNullProperty = property MissingProperty 42
-assert ($MissingNullProperty -eq 42)
+equals $MissingNullProperty 42
 
 # Error: missing property
 task MissingProperty {